+++ app/javascript/app.vue
@@ -1,25 +1,3 @@
 <template>
-  <v-app id="app">
-    <v-btn>Vuetifyのボタン</v-btn>
-    <div>
-      <p>{{ message }}</p>
-    </div>
-  </v-app>
+  <router-view></router-view>
 </template>
-
-<script>
-export default {
-  data: function () {
-    return {
-      message: "Hello Vue!"
-    }
-  }
-}
-</script>
-
-<style scoped>
-p {
-  font-size: 2em;
-  text-align: center;
-}
-</style>
