+++ app/javascript/packs/hello_vue.js
@@ -8,10 +8,12 @@
 import Vue from 'vue'
 import App from '../app.vue'
 import vuetify from './vty'
+import router from '../router/index'
 
 document.addEventListener('DOMContentLoaded', () => {
   const app = new Vue({
     vuetify,
+    router,
     render: h => h(App)
   }).$mount()
   document.body.appendChild(app.$el)
