+++ package.json
@@ -9,6 +9,7 @@
     "turbolinks": "^5.2.0",
     "vue": "^2.6.11",
     "vue-loader": "^15.8.3",
+    "vue-router": "^3.5.2",
     "vue-template-compiler": "^2.6.11",
     "vuetify": "^2.5.8"
   },
